Add macos cwd fix
